+++ CHANGELOG.md
@@ -14,6 +14,7 @@ The format is based on the [KeepAChangeLog] project.
 - [#607] Fixed key recovery on encryption of payload
 - [#618] Prettified `client_management.py` CLI and wrapped it as
          a setup.py console script `oic-client-management`
+- [#615] Fix ROPC grant in the extensions provider
 
 ### Changed
 - [#578] Dropped python 2.7 support
@@ -40,6 +41,7 @@ The format is based on the [KeepAChangeLog] project.
 [#56]: https://github.com/OpenIDC/pyoidc/issues/56
 [#624]: https://github.com/OpenIDC/pyoidc/pull/624
 [#629]: https://github.com/OpenIDC/pyoidc/issues/629
+[#615]: https://github.com/OpenIDC/pyoidc/issues/615
 
 ## 0.15.1 [2019-01-31]